fix: multiple date formats on single page
if multiple ui-date-mask instances exist on a single page they will
share the same dateFormat as this variable is stored in the single
directive instance. E.g. if DD/MM/YYYY and MM/YY are used on the same
page the directive will validate against whichever format was picked up
last.